Renamed feature folder name
